+++ package.json
@@ -8,6 +8,8 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@react-three/drei": "^9.96.5",
+    "@react-three/fiber": "^8.15.15",
     "@reduxjs/toolkit": "^2.0.1",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -21,6 +23,7 @@
     "react-scripts": "5.0.1",
     "sass": "^1.69.3",
     "TagCloud": "^2.4.0",
+    "three": "^0.160.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
