+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import HomePage from './pages/Home';
 import LayoutPage from './pages/Layout';
-import ProjectsPage from './pages/Projects';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import ErrorPage from './pages/Error';
 
@@ -10,13 +9,7 @@ const router = createBrowserRouter([
     path: '/',
     element: <LayoutPage />,
     errorElement: <ErrorPage />,
-    children: [
-      { index: true, element: <HomePage /> },
-      {
-        path: 'projects',
-        element: <ProjectsPage />,
-      },
-    ],
+    children: [{ index: true, element: <HomePage /> }],
   },
 ]);
