--- src/components/Project.module.scss
@@ -1,115 +1,0 @@
-@import '../scss_partials/variables';
-
-.project {
-  display: flex;
-  border-radius: 20px;
-  justify-content: space-between;
-  max-height: 450px;
-  width: 990px;
-  margin: 150px auto;
-  margin-top: 20px;
-  overflow: hidden;
-  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-
-  img {
-    width: 55%;
-    object-fit: cover;
-  }
-
-  @media screen and (min-width: 650px) {
-    &:hover {
-      h1 {
-        color: $secondary-color;
-      }
-    }
-  }
-
-  @media screen and (min-width: 1050px) {
-    &:nth-of-type(2n) {
-      flex-direction: row-reverse;
-    }
-  }
-
-  .info {
-    padding: 40px;
-    text-align: center;
-    align-self: center;
-    padding: 20px;
-
-    .warning {
-      color: orange;
-      font-weight: weight(bold);
-    }
-
-    h1 {
-      transition: 0.7s;
-      margin-bottom: 20px;
-      font-size: 20px;
-    }
-
-    p {
-      font-size: 18px;
-      font-weight: weight(regular);
-      margin-bottom: 40px;
-    }
-
-    .links {
-      display: flex;
-      justify-content: space-around;
-      font-size: 22px;
-      a {
-        text-decoration: none;
-        color: $primary-color;
-        font-weight: weight(medium);
-        transition: 0.5s;
-
-        &:hover {
-          color: $secondary-color;
-        }
-      }
-    }
-  }
-
-  // Responsive
-
-  @media screen and (max-width: 1050px) {
-    padding: 40px;
-    width: 600px;
-    flex-direction: column;
-    align-items: center;
-    max-height: fit-content;
-    margin-top: 0;
-
-    img {
-      width: 95%;
-      margin-bottom: 30px;
-      border-radius: 25px;
-      object-fit: contain;
-    }
-  }
-
-  @media screen and (max-width: 650px) {
-    padding: 30px 10px;
-    width: 100%;
-
-    .info {
-      p {
-        margin-bottom: 10px;
-      }
-
-      .show-description {
-        font-weight: weight(bold);
-        margin: 15px 0;
-        font-size: 19px;
-        background-color: none;
-        border: none;
-        padding: 10px 20px;
-        cursor: pointer;
-      }
-
-      .links {
-        margin-top: 30px;
-      }
-    }
-  }
-}
