+++ src/components/Projects.module.scss
@@ -1,65 +1,132 @@
 @import '../scss_partials/variables';
 
 .projects {
-  max-width: 1024px;
+  @include container;
+
   margin: auto;
   margin-bottom: 300px;
-  color: $primary-color;
+  .info {
+    width: 70%;
+    margin-bottom: 50px;
+  }
+  p {
+    color: $secondary-typo-color;
+    font-size: 20px;
+  }
 
-  .show-projects {
-    display: block;
-    margin: auto;
-    width: fit-content;
-    text-decoration: none;
-    padding: 20px 30px;
-    border: 2px solid $secondary-color;
-    font-weight: weight(bold);
-    color: $secondary-color;
-    border-radius: 20px;
-    transition: 0.5s;
+  .caption {
+    letter-spacing: 0.1rem;
+    margin-bottom: 10px;
+  }
+  .title {
+    font-size: 60px;
+    color: $typography-color;
+    margin-bottom: 20px;
+  }
 
-    @media (min-width: query(small)) {
-      &:hover {
-        background-color: $secondary-color;
-        color: #fff;
+  .disclaimer {
+    color: orangered !important;
+  }
+
+  .projects-section {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    div {
+      width: 320px;
+      padding: 1.25rem;
+      background-color: $secondary-background;
+      border-radius: 20px;
+      margin-bottom: 40px;
+
+      .head {
+        position: relative;
+
+        .source {
+          position: absolute;
+          top: 20px;
+          right: 20px;
+          text-decoration: none;
+          color: $typography-color;
+          font-size: 2.3rem;
+
+          span * {
+            background-color: #fff;
+            border-radius: 50%;
+            border: 1px solid #fff;
+          }
+        }
+        .link {
+          position: absolute;
+          top: 20px;
+          left: 20px;
+          text-decoration: none;
+          color: $accent-blue;
+          font-size: 2rem;
+        }
+      }
+
+      h1 {
+        font-size: 24px;
+        background: linear-gradient(to bottom, $primary-blue, $accent-blue);
+        background-clip: text;
+        -webkit-background-clip: text;
+        color: transparent;
+        margin-bottom: 10px;
+      }
+
+      p {
+        font-size: 14px;
+        color: $secondary-typo-color;
+        margin-bottom: 10px;
+      }
+
+      ul {
+        list-style: none;
+        display: flex;
+        flex-wrap: wrap;
+        li {
+          margin-right: 10px;
+          margin-bottom: 10px;
+        }
+      }
+
+      img {
+        width: 100%;
+        height: 230px;
+        object-fit: cover;
+        border-radius: 20px;
+        margin-bottom: 20px;
       }
     }
   }
 
-  span {
-    display: inline-block;
-    color: $secondary-color;
-    font-weight: weight(bold);
-    font-size: 18px;
-    margin-bottom: 15px;
-  }
+  @media screen and (max-width: 1024px) {
+    .projects-section {
+      justify-content: center;
+      div {
+        margin-right: 50px;
+      }
 
-  h1 {
-    font-size: 25px;
-    font-weight: weight(bold);
-    margin-bottom: 100px;
-  }
-
-  // Responsive
-
-  @media screen and (max-width: 1050px) {
-    max-width: fit-content;
-    margin: auto;
-    margin-bottom: 200px;
-
-    h1,
-    span {
-      font-size: 27px;
+      @include mobile {
+        div {
+          margin-right: 0;
+          width: 80%;
+        }
+      }
     }
   }
 
-  @media screen and (max-width: 650px) {
-    width: 95%;
-    text-align: center;
-
-    h1 {
-      font-size: 23px;
-      margin-bottom: 50px;
+  @include mobile {
+    .info {
+      width: 100%;
+      padding: 0 30px;
+    }
+    .caption {
+      font-size: 18px;
+    }
+    .title {
+      font-size: 40px;
     }
   }
 }
