--- src/components/Projects2.module.scss
@@ -1,128 +1,0 @@
-@import '../scss_partials/variables';
-
-.projects {
-  @include container;
-
-  margin: auto;
-  margin-bottom: 300px;
-  .info {
-    width: 70%;
-    margin-bottom: 50px;
-  }
-  p {
-    color: $secondary-typo-color;
-    font-size: 20px;
-  }
-
-  .caption {
-    letter-spacing: 0.1rem;
-    margin-bottom: 10px;
-  }
-  .title {
-    font-size: 60px;
-    color: $typography-color;
-    margin-bottom: 20px;
-  }
-
-  .projects-section {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    div {
-      width: 320px;
-      padding: 1.25rem;
-      background-color: $secondary-background;
-      border-radius: 20px;
-      margin-bottom: 40px;
-
-      .head {
-        position: relative;
-
-        .source {
-          position: absolute;
-          top: 20px;
-          right: 20px;
-          text-decoration: none;
-          color: $typography-color;
-          font-size: 2.3rem;
-
-          span * {
-            background-color: #fff;
-            border-radius: 50%;
-            border: 1px solid #fff;
-          }
-        }
-        .link {
-          position: absolute;
-          top: 20px;
-          left: 20px;
-          text-decoration: none;
-          color: $accent-blue;
-          font-size: 2rem;
-        }
-      }
-
-      h1 {
-        font-size: 24px;
-        background: linear-gradient(to bottom, $primary-blue, $accent-blue);
-        background-clip: text;
-        -webkit-background-clip: text;
-        color: transparent;
-        margin-bottom: 10px;
-      }
-
-      p {
-        font-size: 14px;
-        color: $secondary-typo-color;
-        margin-bottom: 10px;
-      }
-
-      ul {
-        list-style: none;
-        display: flex;
-        flex-wrap: wrap;
-        li {
-          margin-right: 10px;
-          margin-bottom: 10px;
-        }
-      }
-
-      img {
-        width: 100%;
-        height: 230px;
-        object-fit: cover;
-        border-radius: 20px;
-        margin-bottom: 20px;
-      }
-    }
-  }
-
-  @media screen and (max-width: 1024px) {
-    .projects-section {
-      justify-content: center;
-      div {
-        margin-right: 50px;
-      }
-
-      @include mobile {
-        div {
-          margin-right: 0;
-          width: 80%;
-        }
-      }
-    }
-  }
-
-  @include mobile {
-    .info {
-      width: 100%;
-      padding: 0 30px;
-    }
-    .caption {
-      font-size: 18px;
-    }
-    .title {
-      font-size: 40px;
-    }
-  }
-}
