+++ src/pages/Home.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import Hero from '../components/Hero';
 import AboutMe from '../components/AboutMe';
-// import Projects from '../components/Projects';
 import DevTree from '../components/DevTree';
 import TechStack from '../components/TechStack';
 
 import { skills } from '../assets/SKILLS';
 import { PROJECTS } from '../assets/PROJECTS';
-import Projects2 from '../components/Projects2';
+import Projects from '../components/Projects';
 
 function MainPage() {
   return (
@@ -16,8 +15,7 @@ function MainPage() {
       <AboutMe />
       <DevTree />
       <TechStack techStack={skills} />
-      <Projects2 projects={PROJECTS} />
-      {/* <Projects /> */}
+      <Projects projects={PROJECTS} />
     </>
   );
 }
