--- src/pages/Projects.module.scss
@@ -1,71 +1,0 @@
-@import '../scss_partials/variables';
-
-.header {
-  margin: auto;
-  width: fit-content;
-  margin-top: 300px;
-  margin-bottom: 50px;
-  font-size: 30px;
-  color: $primary-color;
-  font-weight: weight(bold);
-}
-.projects {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  width: 1200px;
-  margin: auto;
-
-  .project {
-    color: $primary-color;
-    padding: 20px;
-    border: 3px solid $secondary-color;
-    border-radius: 20px;
-    text-align: center;
-    margin: 20px;
-
-    @media screen and (min-width: 1300px) {
-      width: 30%;
-    }
-
-    h1 {
-      margin-bottom: 30px;
-      border-bottom: 3px solid $secondary-color;
-      padding-bottom: 20px;
-    }
-
-    .icon {
-      margin: 0 10px;
-      font-size: 1.8rem;
-
-      a {
-        text-decoration: none;
-        color: $primary-color;
-        transition: 0.5s;
-
-        @media screen and (min-width: query(small)) {
-          &:hover {
-            color: $secondary-color;
-          }
-        }
-      }
-    }
-
-    img {
-      width: 300px;
-      height: 200px;
-      margin-bottom: 20px;
-      border-radius: 20px;
-    }
-  }
-
-  @media screen and (max-width: 1240px) {
-    width: 90%;
-
-    img {
-      width: 200px;
-      height: 150px;
-    }
-  }
-}
